# backend/api_doc/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is an API that serve services to an Online store, It is able to manage users, shopping carts, products loader, etc.",
    "version": "0.1.0",
    "title": "Online store"
  },
  "host": "kod3rsstore.com",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/addresses": {
      "post": {
        "tags": [
          "Addresses"
        ],
        "summary": "Add a new address to an user",
        "description": "",
        "operationId": "addAddress",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Address object to be saved into user account",
            "required": true,
            "schema": {
              "$ref": "#/definitions/createAddresses"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "address created"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": [
          "Addresses"
        ],
        "summary": "update the address information",
        "description": "You will be able to change the addres of an user",
        "operationId": "updateAddresses",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Information of address to be updated",
            "required": true,
            "schema": {
              "$ref": "#/definitions/updateAddresses"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "address updated"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/addresses/{id}": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "summary": "Gets all addresses of an user, by user ID",
        "description": "You might fetch all addresses of an user",
        "operationId": "getAddressById",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ID of an authorized user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "get the addresses",
            "schema": {
              "$ref": "#/definitions/responseGetAddresses"
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "createAddresses": {
      "type": "object",
      "required": [
        "id_user",
        "id_cities_catalog",
        "id_countries_catalog",
        "street"
      ],
      "properties": {
        "id_user": {
          "type": "string",
          "example": "Ah81pi_zHZreS8yvL3rMc"
        },
        "id_cities_catalog": {
          "type": "string",
          "example": "2"
        },
        "id_countries_catalog": {
          "type": "string",
          "example": "2"
        },
        "id_states_catalog": {
          "type": "string",
          "example": "2"
        },
        "street": {
          "type": "string",
          "example": "carrera 20 # 12-25 oficina 1203"
        },
        "street_number": {
          "type": "string",
          "example": "La 20"
        },
        "id_postal_codes": {
          "type": "string",
          "example": "112356"
        }
      }
    },
    "updateAddresses": {
      "type": "object",
      "required": [
        "id_directions",
        "id_user"
      ],
      "properties": {
        "id_user": {
          "type": "string",
          "example": "Ah81pi_zHZreS8yvL3rMc"
        },
        "id_directions": {
          "type": "string",
          "example": "l7HHxanQbEDZDU6myrPTU"
        },
        "id_cities_catalog": {
          "type": "string",
          "example": "2"
        },
        "id_countries_catalog": {
          "type": "string",
          "example": "2"
        },
        "id_states_catalog": {
          "type": "string",
          "example": "2"
        },
        "street": {
          "type": "string",
          "example": "carrera 20 # 12-25 oficina 1203"
        },
        "street_number": {
          "type": "string",
          "example": "La 20"
        },
        "id_postal_codes": {
          "type": "string",
          "example": "112356"
        },
        "available": {
          "type": "boolean",
          "example": true
        }
      }
    },
...
    "responseGetAddresses": {
      "type": "object",
      "properties": {
        "error": {
          "type": "boolean",
          "example": false
        },
        "status": {
          "type": "integer",
          "example": 200
        },
        "body": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id_directions": {
                "type": "string",
                "example": "F-4fVj1__ZZP-M9wmFNUS"
              },
              "id_cities_catalog": {
                "type": "string",
                "example": "1"
              },
              "id_countries_catalog": {
                "type": "string",
                "example": "1"
              },
              "id_states_catalog": {
                "type": "string",
                "example": "1"
              },
              "street": {
                "type": "string",
                "example": "Carrera 10 n 34-50"
              },
              "street_number": {
                "type": "string",
                "example": "la 20"
              },
              "id_postal_codes": {
                "type": "string",
                "example": "11032388"
              },
              "available": {
                "type": "boolean",
                "example": true
              },
              "creation_date": {
                "type": "string",
                "example": "2020-05-17T05:00:00.000Z"
              },
              "id_users": {
                "type": "string",
                "example": "0kfJIxZCKeRopOjtx4fVg"
              }
            }
          }
        }
      }
    },
...
  },
...
}